move print functions to separate file
